+++ packages/web-component/src/lib/descope-wc/BaseDescopeWc.ts
@@ -146,6 +146,35 @@ class BaseDescopeWc extends HTMLElement {
     return this.getAttribute('flow-id');
   }
 
+  get form() {
+    try {
+      const form = (JSON.parse(this.getAttribute('form')) || {}) as Record<
+        string,
+        any
+      >;
+      return Object.entries(form).reduce(
+        (prev, [key, value]) => ({
+          ...prev,
+          [`form.${key}`]: value,
+        }),
+        form,
+      );
+    } catch (e) {
+      return {};
+    }
+  }
+
+  get client() {
+    try {
+      return (JSON.parse(this.getAttribute('client')) || {}) as Record<
+        string,
+        any
+      >;
+    } catch (e) {
+      return {};
+    }
+  }
+
   get baseUrl() {
     return this.getAttribute('base-url') || undefined;
   }
@@ -206,6 +235,8 @@ class BaseDescopeWc extends HTMLElement {
       'auto-focus',
       'preview',
       'storage-prefix',
+      'form',
+      'client',
     ];
 
     BaseDescopeWc.observedAttributes.forEach((attr: string) => {
@@ -215,7 +246,7 @@ class BaseDescopeWc extends HTMLElement {
 
     if (this.theme && this.theme !== 'light' && this.theme !== 'dark') {
       throw Error(
-        'Supported theme values are "light", "dark", or leave empty for using the OS theme'
+        'Supported theme values are "light", "dark", or leave empty for using the OS theme',
       );
     }
   }
@@ -265,7 +296,7 @@ class BaseDescopeWc extends HTMLElement {
   async #onFlowChange(
     currentState: FlowState,
     _prevState: FlowState,
-    isChanged: IsChanged<FlowState>
+    isChanged: IsChanged<FlowState>,
   ) {
     const { projectId, baseUrl } = currentState;
 
@@ -292,7 +323,7 @@ class BaseDescopeWc extends HTMLElement {
     const prevVerConfigUrl = getContentUrl(
       this.projectId,
       CONFIG_FILENAME,
-      PREV_VER_ASSETS_FOLDER
+      PREV_VER_ASSETS_FOLDER,
     );
     try {
       await fetchContent(prevVerConfigUrl, 'json');
@@ -322,7 +353,7 @@ class BaseDescopeWc extends HTMLElement {
     // eslint-disable-next-line @typescript-eslint/no-unused-expressions
     fonts &&
       Object.values(fonts).forEach((font: Record<string, any>) =>
-        loadFont(font.url)
+        loadFont(font.url),
       );
   }
 
@@ -349,7 +380,7 @@ class BaseDescopeWc extends HTMLElement {
     } catch (e) {
       this.loggerWrapper.error(
         'Cannot fetch theme file',
-        'make sure that your projectId & flowId are correct'
+        'make sure that your projectId & flowId are correct',
       );
     }
     this.shadowRoot.appendChild(styleEle);
@@ -381,7 +412,7 @@ class BaseDescopeWc extends HTMLElement {
   async #handleDebugMode({ isDebug }) {
     if (isDebug) {
       this.#debuggerEle = document.createElement(
-        'descope-debugger'
+        'descope-debugger',
       ) as HTMLElement & {
         updateData: (data: DebuggerMessage | DebuggerMessage[]) => void;
       };
@@ -427,7 +458,7 @@ class BaseDescopeWc extends HTMLElement {
   async getTargetLocales() {
     const flowConfig = await this.getFlowConfig();
     return (flowConfig?.targetLocales || []).map((locale: string) =>
-      locale.toLowerCase()
+      locale.toLowerCase(),
     );
   }
 
@@ -447,7 +478,7 @@ class BaseDescopeWc extends HTMLElement {
       }
 
       const genericButtons = Array.from(buttons).filter(
-        (button) => button.getAttribute('data-type') === 'button'
+        (button) => button.getAttribute('data-type') === 'button',
       );
 
       // in case there is a single "generic" button on the page, click on it
@@ -467,14 +498,14 @@ class BaseDescopeWc extends HTMLElement {
     }
 
     const existingScript = document.getElementById(
-      SCRIPT_ID
+      SCRIPT_ID,
     ) as HTMLScriptElement;
     const scriptEle: HTMLScriptElement =
       existingScript || document.createElement('script');
 
     if (existingScript) {
       this.loggerWrapper.info(
-        'DescopeUI loading script is already exist, probably multiple flows are running on the same page'
+        'DescopeUI loading script is already exist, probably multiple flows are running on the same page',
       );
     } else {
       scriptEle.id = SCRIPT_ID;
@@ -483,14 +514,14 @@ class BaseDescopeWc extends HTMLElement {
 
       if (!version) {
         this.logger.error(
-          'Did not get components version, using latest version'
+          'Did not get components version, using latest version',
         );
         version = 'latest';
       }
 
       scriptEle.src = UI_COMPONENTS_URL.replace(
         UI_COMPONENTS_URL_VERSION_PLACEHOLDER,
-        version
+        version,
       );
 
       document.body.append(scriptEle);
@@ -500,7 +531,7 @@ class BaseDescopeWc extends HTMLElement {
       const onError = () => {
         this.loggerWrapper.error(
           'Cannot load DescopeUI',
-          `Make sure this URL is valid and return the correct script: "${scriptEle.src}"`
+          `Make sure this URL is valid and return the correct script: "${scriptEle.src}"`,
         );
 
         res(undefined);
@@ -535,7 +566,7 @@ class BaseDescopeWc extends HTMLElement {
       if (await this.#getIsFlowsVersionMismatch()) {
         this.loggerWrapper.error(
           'This SDK version does not support your flows version',
-          'Make sure to upgrade your flows to the latest version or use an older SDK version'
+          'Make sure to upgrade your flows to the latest version or use an older SDK version',
         );
 
         return;
@@ -544,7 +575,7 @@ class BaseDescopeWc extends HTMLElement {
       if ((await this.#getConfig()).isMissingConfig) {
         this.loggerWrapper.error(
           'Cannot get config file',
-          'Make sure that your projectId & flowId are correct'
+          'Make sure that your projectId & flowId are correct',
         );
 
         return;
@@ -580,12 +611,12 @@ class BaseDescopeWc extends HTMLElement {
       // this data will be sent to the server on the next request
       window.addEventListener(
         'components-context',
-        this.#eventsCbRefs.componentsContext
+        this.#eventsCbRefs.componentsContext,
       );
 
       window.addEventListener(
         'visibilitychange',
-        this.#eventsCbRefs.visibilitychange
+        this.#eventsCbRefs.visibilitychange,
       );
 
       this.#flowState.subscribe(this.#onFlowChange.bind(this));
@@ -622,18 +653,18 @@ class BaseDescopeWc extends HTMLElement {
     window.removeEventListener('popstate', this.#eventsCbRefs.popstate);
     window.removeEventListener(
       'visibilitychange',
-      this.#eventsCbRefs.visibilitychange
+      this.#eventsCbRefs.visibilitychange,
     );
     window.removeEventListener(
       'components-context',
-      this.#eventsCbRefs.componentsContext
+      this.#eventsCbRefs.componentsContext,
     );
   }
 
   attributeChangedCallback(
     attrName: string,
     oldValue: string,
-    newValue: string
+    newValue: string,
   ) {
     if (!this.shadowRoot.isConnected || !this.#init) return;