+++ packages/web-component/test/descope-wc.test.ts
@@ -195,7 +195,7 @@ describe('web-component', () => {
         ok: false,
         requestErrorMessage: 'Not found',
         requestErrorDescription: 'Not found',
-      })
+      }),
     );
 
     document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="otpSignInEmail" project-id="1"></descope-wc>`;
@@ -213,9 +213,9 @@ describe('web-component', () => {
               errorMessage: 'Not found',
               errorDescription: 'Not found',
             },
-          })
+          }),
         ),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
 
     wcEle.removeEventListener('error', onError);
@@ -250,13 +250,13 @@ describe('web-component', () => {
     });
 
     const themeStyleEle = shadowEle?.querySelector(
-      'style:last-child'
+      'style:last-child',
     ) as HTMLStyleElement;
     expect(themeStyleEle.innerText).toContain(
-      (themeContent as any).light.globals
+      (themeContent as any).light.globals,
     );
     expect(themeStyleEle.innerText).toContain(
-      (themeContent as any).dark.globals
+      (themeContent as any).dark.globals,
     );
   });
 
@@ -301,7 +301,7 @@ describe('web-component', () => {
     expect(autoFocusSpy).toBeCalledWith(
       expect.any(HTMLElement),
       'skipFirstScreen',
-      true
+      true,
     );
     autoFocusSpy.mockClear();
 
@@ -311,10 +311,10 @@ describe('web-component', () => {
         expect(autoFocusSpy).toBeCalledWith(
           expect.any(HTMLElement),
           'skipFirstScreen',
-          false
+          false,
         );
       },
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
@@ -339,17 +339,17 @@ describe('web-component', () => {
 
     expect(fetchMock).toHaveBeenCalledWith(
       expect.stringMatching(htmlUrlPathRegex),
-      expect.any(Object)
+      expect.any(Object),
     );
 
     expect(fetchMock).toHaveBeenCalledWith(
       expect.stringMatching(themeUrlPathRegex),
-      expect.any(Object)
+      expect.any(Object),
     );
 
     expect(fetchMock).toHaveBeenCalledWith(
       expect.stringMatching(configUrlPathRegex),
-      expect.any(Object)
+      expect.any(Object),
     );
   });
 
@@ -376,7 +376,7 @@ describe('web-component', () => {
     });
 
     await expect(descope.connectedCallback.bind(descope)).rejects.toThrow(
-      'project-id cannot be empty'
+      'project-id cannot be empty',
     );
   });
 
@@ -402,7 +402,7 @@ describe('web-component', () => {
     });
 
     await expect(descope.connectedCallback.bind(descope)).rejects.toThrow(
-      'flow-id cannot be empty'
+      'flow-id cannot be empty',
     );
   });
 
@@ -478,8 +478,8 @@ describe('web-component', () => {
         {
           email: '',
           origin: 'http://localhost',
-        }
-      )
+        },
+      ),
     );
   });
 
@@ -508,8 +508,8 @@ describe('web-component', () => {
         {
           t1: false,
           origin: 'http://localhost',
-        }
-      )
+        },
+      ),
     );
   });
 
@@ -538,8 +538,8 @@ describe('web-component', () => {
         {
           t1: true,
           origin: 'http://localhost',
-        }
-      )
+        },
+      ),
     );
   });
 
@@ -574,6 +574,7 @@ describe('web-component', () => {
           samlIdpStateId: null,
           samlIdpUsername: null,
           ssoAppId: null,
+          client: {},
           redirectAuth: undefined,
           tenant: undefined,
         },
@@ -585,8 +586,8 @@ describe('web-component', () => {
           email: '',
           origin: 'http://localhost',
           token,
-        }
-      )
+        },
+      ),
     );
   });
 
@@ -638,8 +639,8 @@ describe('web-component', () => {
         'click',
         1,
         '1.2.3',
-        expect.any(Object)
-      )
+        expect.any(Object),
+      ),
     );
   });
 
@@ -690,7 +691,7 @@ describe('web-component', () => {
   it('should update the page messages when page is remaining the same but the state is updated', async () => {
     startMock.mockReturnValueOnce(generateSdkResponse());
     nextMock.mockReturnValueOnce(
-      generateSdkResponse({ screenState: { errorText: 'Error!' } })
+      generateSdkResponse({ screenState: { errorText: 'Error!' } }),
     );
 
     pageContent = `<descope-button>click</descope-button><div>Loaded1</div><span ${ELEMENT_TYPE_ATTRIBUTE}="error-message">xxx</span>`;
@@ -710,14 +711,14 @@ describe('web-component', () => {
         screen.getByShadowText('Error!', {
           selector: `[${ELEMENT_TYPE_ATTRIBUTE}="error-message"]`,
         }),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
   it('should update page inputs according to screen state', async () => {
     startMock.mockReturnValueOnce(generateSdkResponse());
     nextMock.mockReturnValueOnce(
-      generateSdkResponse({ screenState: { inputs: { email: 'email1' } } })
+      generateSdkResponse({ screenState: { inputs: { email: 'email1' } } }),
     );
 
     pageContent = `<descope-button>click</descope-button><div>Loaded</div><input class="descope-input" name="email">`;
@@ -763,7 +764,7 @@ describe('web-component', () => {
           image: 'data:;base64,example',
           origin: 'http://localhost',
         }),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
@@ -790,13 +791,13 @@ describe('web-component', () => {
           image: '',
           origin: 'http://localhost',
         }),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
   it('should update page templates according to screen state', async () => {
     startMock.mockReturnValue(
-      generateSdkResponse({ screenState: { user: { name: 'john' } } })
+      generateSdkResponse({ screenState: { user: { name: 'john' } } }),
     );
 
     pageContent = `<div>Loaded1</div><descope-text class="descope-text">hey {{user.name}}!</descope-text>`;
@@ -811,7 +812,7 @@ describe('web-component', () => {
 
   it('should update page templates according to last auth login ID when there is only login Id', async () => {
     startMock.mockReturnValue(
-      generateSdkResponse({ screenState: { user: { name: 'john' } } })
+      generateSdkResponse({ screenState: { user: { name: 'john' } } }),
     );
     getLastUserLoginIdMock.mockReturnValue('not john');
 
@@ -827,7 +828,7 @@ describe('web-component', () => {
 
   it('should update page templates according to last auth name when there is only login Id', async () => {
     startMock.mockReturnValue(
-      generateSdkResponse({ screenState: { user: { name: 'john' } } })
+      generateSdkResponse({ screenState: { user: { name: 'john' } } }),
     );
     getLastUserLoginIdMock.mockReturnValue('not john');
 
@@ -843,7 +844,7 @@ describe('web-component', () => {
 
   it('should update page templates according to last auth name when there is login Id and name', async () => {
     startMock.mockReturnValue(
-      generateSdkResponse({ screenState: { user: { name: 'john' } } })
+      generateSdkResponse({ screenState: { user: { name: 'john' } } }),
     );
     getLastUserLoginIdMock.mockReturnValue('not john');
     getLastUserDisplayNameMock.mockReturnValue('Niros!');
@@ -862,7 +863,7 @@ describe('web-component', () => {
 
   it('should update totp link href according to screen state', async () => {
     startMock.mockReturnValue(
-      generateSdkResponse({ screenState: { totp: { provisionUrl: 'url1' } } })
+      generateSdkResponse({ screenState: { totp: { provisionUrl: 'url1' } } }),
     );
 
     pageContent = `<div>Loaded1</div><descope-link data-type="totp-link">Provision URL</descope-link>`;
@@ -897,7 +898,7 @@ describe('web-component', () => {
 
   it('should update root css var according to screen state', async () => {
     startMock.mockReturnValue(
-      generateSdkResponse({ screenState: { totp: { image: 'base-64-text' } } })
+      generateSdkResponse({ screenState: { totp: { image: 'base-64-text' } } }),
     );
 
     const spyGet = jest.spyOn(customElements, 'get');
@@ -919,7 +920,7 @@ describe('web-component', () => {
         expect(rootEle).toHaveStyle({
           '--url': 'url(data:image/jpg;base64,base-64-text)',
         }),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
@@ -943,7 +944,7 @@ describe('web-component', () => {
     fireEvent.popState(window);
 
     await waitFor(() =>
-      expect(logSpy).toHaveBeenCalledWith('No screen was found to show', '')
+      expect(logSpy).toHaveBeenCalledWith('No screen was found to show', ''),
     );
   });
 
@@ -968,9 +969,9 @@ describe('web-component', () => {
         expect(errorSpy).toHaveBeenCalledWith(
           'Cannot get config file',
           'Make sure that your projectId & flowId are correct',
-          expect.any(Error)
+          expect.any(Error),
         ),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
@@ -996,7 +997,7 @@ describe('web-component', () => {
     const spyAddEventListener = jest.spyOn(rootEle, 'addEventListener');
 
     spyAddEventListener.mockImplementationOnce(
-      (_, cb) => typeof cb === 'function' && cb({} as Event)
+      (_, cb) => typeof cb === 'function' && cb({} as Event),
     );
 
     await waitFor(() => screen.findByShadowText('It updated!'), {
@@ -1015,7 +1016,7 @@ describe('web-component', () => {
     await waitFor(() =>
       expect(nextMock).toHaveBeenCalledWith('0', '0', 'submit', 1, '1.2.3', {
         token: 'token1',
-      })
+      }),
     );
     await waitFor(() => screen.findByShadowText('It works!'), {
       timeout: WAIT_TIMEOUT,
@@ -1033,7 +1034,7 @@ describe('web-component', () => {
     await waitFor(() =>
       expect(nextMock).toHaveBeenCalledWith('0', '0', 'submit', 0, '1.2.3', {
         exchangeCode: 'code1',
-      })
+      }),
     );
     await waitFor(() => screen.findByShadowText('It works!'), {
       timeout: WAIT_TIMEOUT,
@@ -1051,7 +1052,7 @@ describe('web-component', () => {
     await waitFor(() =>
       expect(nextMock).toHaveBeenCalledWith('0', '0', 'submit', 1, '1.2.3', {
         exchangeError: 'err1',
-      })
+      }),
     );
     await waitFor(() => screen.findByShadowText('It works!'), {
       timeout: WAIT_TIMEOUT,
@@ -1080,7 +1081,7 @@ describe('web-component', () => {
         attr1: 'attr1',
         attr2: 'attr2',
         origin: 'http://localhost',
-      })
+      }),
     );
   });
 
@@ -1091,7 +1092,7 @@ describe('web-component', () => {
       () =>
         new Promise((res) => {
           resolve = res;
-        })
+        }),
     );
 
     pageContent = `<descope-button type="button" id="123" ${DESCOPE_ATTRIBUTE_PREFIX}attr1='attr1' ${DESCOPE_ATTRIBUTE_PREFIX}attr2='attr2'>Click</descope-button>`;
@@ -1105,14 +1106,17 @@ describe('web-component', () => {
     fireEvent.click(screen.getByShadowText('Click'));
 
     await waitFor(() =>
-      expect(screen.getByShadowText('Click')).toHaveAttribute('loading', 'true')
+      expect(screen.getByShadowText('Click')).toHaveAttribute(
+        'loading',
+        'true',
+      ),
     );
 
     resolve(generateSdkResponse({ screenId: '1' }));
 
     await waitFor(
       () => expect(screen.getByShadowText('Click')).not.toHaveClass('loading'),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
@@ -1121,7 +1125,7 @@ describe('web-component', () => {
       generateSdkResponse({
         action: RESPONSE_ACTIONS.redirect,
         redirectUrl: 'https://myurl.com',
-      })
+      }),
     );
 
     window.location.search = `?${URL_RUN_IDS_PARAM_NAME}=0_0&${URL_CODE_PARAM_NAME}=code1`;
@@ -1130,11 +1134,11 @@ describe('web-component', () => {
     await waitFor(
       () =>
         expect(window.location.assign).toHaveBeenCalledWith(
-          'https://myurl.com'
+          'https://myurl.com',
         ),
       {
         timeout: WAIT_TIMEOUT,
-      }
+      },
     );
   });
 
@@ -1143,7 +1147,7 @@ describe('web-component', () => {
       generateSdkResponse({
         action: RESPONSE_ACTIONS.redirect,
         redirectUrl: 'https://myurl.com',
-      })
+      }),
     );
 
     window.location.search = `?${URL_RUN_IDS_PARAM_NAME}=0_0&${URL_CODE_PARAM_NAME}=code1`;
@@ -1153,7 +1157,7 @@ describe('web-component', () => {
       () => expect(window.location.assign).toHaveBeenCalledTimes(1),
       {
         timeout: WAIT_TIMEOUT,
-      }
+      },
     );
   });
 
@@ -1161,7 +1165,7 @@ describe('web-component', () => {
     startMock.mockReturnValueOnce(
       generateSdkResponse({
         action: RESPONSE_ACTIONS.redirect,
-      })
+      }),
     );
 
     const errorSpy = jest.spyOn(console, 'error');
@@ -1173,9 +1177,9 @@ describe('web-component', () => {
         expect(errorSpy).toHaveBeenCalledWith(
           'Did not get redirect url',
           '',
-          expect.any(Error)
+          expect.any(Error),
         ),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
@@ -1184,7 +1188,7 @@ describe('web-component', () => {
       generateSdkResponse({
         action: RESPONSE_ACTIONS.redirect,
         redirectUrl: 'https://myurl.com',
-      })
+      }),
     );
 
     // Start hidden (in background)
@@ -1204,7 +1208,7 @@ describe('web-component', () => {
       timeout: WAIT_TIMEOUT,
     });
     expect(window.location.assign).not.toHaveBeenCalledWith(
-      'https://myurl.com'
+      'https://myurl.com',
     );
 
     // Back to the foreground
@@ -1213,11 +1217,11 @@ describe('web-component', () => {
     await waitFor(
       () =>
         expect(window.location.assign).toHaveBeenCalledWith(
-          'https://myurl.com'
+          'https://myurl.com',
         ),
       {
         timeout: WAIT_TIMEOUT,
-      }
+      },
     );
   });
 
@@ -1226,7 +1230,7 @@ describe('web-component', () => {
       generateSdkResponse({
         action: RESPONSE_ACTIONS.redirect,
         redirectUrl: 'https://myurl.com',
-      })
+      }),
     );
 
     // Start hidden (in background)
@@ -1248,11 +1252,11 @@ describe('web-component', () => {
     await waitFor(
       () =>
         expect(window.location.assign).toHaveBeenCalledWith(
-          'https://myurl.com'
+          'https://myurl.com',
         ),
       {
         timeout: WAIT_TIMEOUT,
-      }
+      },
     );
   });
 
@@ -1260,7 +1264,7 @@ describe('web-component', () => {
     startMock.mockReturnValueOnce(
       generateSdkResponse({
         action: RESPONSE_ACTIONS.webauthnCreate,
-      })
+      }),
     );
 
     const errorSpy = jest.spyOn(console, 'error');
@@ -1272,9 +1276,9 @@ describe('web-component', () => {
         expect(errorSpy).toHaveBeenCalledWith(
           'Did not get webauthn transaction id or options',
           '',
-          expect.any(Error)
+          expect.any(Error),
         ),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
   });
 
@@ -1284,21 +1288,21 @@ describe('web-component', () => {
         action: RESPONSE_ACTIONS.webauthnCreate,
         webAuthnTransactionId: 't1',
         webAuthnOptions: 'options',
-      })
+      }),
     );
     pageContent = '<span>It works!</span>';
 
     nextMock.mockReturnValueOnce(generateSdkResponse());
 
     sdk.webauthn.helpers.create.mockReturnValueOnce(
-      Promise.resolve('webauthn-response')
+      Promise.resolve('webauthn-response'),
     );
 
     document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="webauthn_signup" project-id="1"></descope-wc>`;
 
     await waitFor(
       () => expect(sdk.webauthn.helpers.create).toHaveBeenCalled(),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
     expect(nextMock).toHaveBeenCalledWith('0', '0', 'submit', 0, '1.2.3', {
       transactionId: 't1',
@@ -1312,7 +1316,7 @@ describe('web-component', () => {
         action: RESPONSE_ACTIONS.webauthnGet,
         webAuthnTransactionId: 't1',
         webAuthnOptions: 'options',
-      })
+      }),
     );
 
     pageContent = '<span>It works!</span>';
@@ -1320,7 +1324,7 @@ describe('web-component', () => {
     nextMock.mockReturnValueOnce(generateSdkResponse());
 
     sdk.webauthn.helpers.get.mockReturnValueOnce(
-      Promise.resolve('webauthn-response-get')
+      Promise.resolve('webauthn-response-get'),
     );
 
     document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="versioned-flow" project-id="1"></descope-wc>`;
@@ -1340,7 +1344,7 @@ describe('web-component', () => {
         action: RESPONSE_ACTIONS.webauthnGet,
         webAuthnTransactionId: 't1',
         webAuthnOptions: 'options',
-      })
+      }),
     );
 
     pageContent = '<span>It works!</span>';
@@ -1348,7 +1352,7 @@ describe('web-component', () => {
     nextMock.mockReturnValueOnce(generateSdkResponse());
 
     sdk.webauthn.helpers.get.mockReturnValueOnce(
-      Promise.reject(new DOMException('', 'NotAllowedError'))
+      Promise.reject(new DOMException('', 'NotAllowedError')),
     );
 
     document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="webauthn_signup" project-id="1"></descope-wc>`;
@@ -1376,8 +1380,8 @@ describe('web-component', () => {
 
     await waitFor(() =>
       expect(
-        document.head.querySelector(`link[href="font.url"]`)
-      ).toBeInTheDocument()
+        document.head.querySelector(`link[href="font.url"]`),
+      ).toBeInTheDocument(),
     );
   });
 
@@ -1403,7 +1407,7 @@ describe('web-component', () => {
 
     expect(fetchMock).toHaveBeenCalledWith(
       expect.stringMatching(htmlUrlPathRegex),
-      expect.any(Object)
+      expect.any(Object),
     );
   });
 
@@ -1499,7 +1503,7 @@ describe('web-component', () => {
     });
 
     await expect(descope.connectedCallback.bind(descope)).rejects.toThrow(
-      `Supported theme values are "light", "dark", or leave empty for using the OS theme`
+      `Supported theme values are "light", "dark", or leave empty for using the OS theme`,
     );
   });
 
@@ -1518,7 +1522,7 @@ describe('web-component', () => {
     const buttonEle = await screen.findByShadowText('click');
 
     const inputEle = screen.getByShadowPlaceholderText(
-      'email'
+      'email',
     ) as HTMLInputElement;
 
     inputEle.reportValidity = jest.fn();
@@ -1553,8 +1557,34 @@ describe('web-component', () => {
         '',
         0,
         '1.2.3',
-        undefined
-      )
+        {},
+      ),
+    );
+  });
+
+  it('should call start with form and client when provided', async () => {
+    startMock.mockReturnValueOnce(generateSdkResponse());
+
+    pageContent = '<div>hey</div>';
+
+    document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="sign-in" project-id="1" form='{"email": "test"}' client='{"email": "test2"}'></descope-wc>`;
+
+    await waitFor(() => screen.findByShadowText('hey'), {
+      timeout: WAIT_TIMEOUT,
+    });
+
+    await waitFor(() =>
+      expect(startMock).toHaveBeenCalledWith(
+        'sign-in',
+        expect.objectContaining({
+          client: { email: 'test2' },
+        }),
+        undefined,
+        '',
+        0,
+        '1.2.3',
+        { email: 'test', 'form.email': 'test' },
+      ),
     );
   });
 
@@ -1576,7 +1606,7 @@ describe('web-component', () => {
           stepId: 's1',
           screenId: '1',
           action: RESPONSE_ACTIONS.poll,
-        })
+        }),
       );
 
       nextMock.mockReturnValueOnce(generateSdkResponse());
@@ -1600,7 +1630,7 @@ describe('web-component', () => {
       nextMock.mockReturnValueOnce(
         generateSdkResponse({
           action: RESPONSE_ACTIONS.poll,
-        })
+        }),
       );
 
       pageContent = '<div data-type="polling">...</div><span>It works!</span>';
@@ -1613,7 +1643,7 @@ describe('web-component', () => {
           expect(setInterval).toHaveBeenCalledWith(expect.any(Function), 2000),
         {
           timeout: WAIT_TIMEOUT,
-        }
+        },
       );
     });
 
@@ -1623,7 +1653,7 @@ describe('web-component', () => {
       nextMock.mockReturnValueOnce(
         generateSdkResponse({
           action: RESPONSE_ACTIONS.poll,
-        })
+        }),
       );
 
       pageContent =
@@ -1639,11 +1669,11 @@ describe('web-component', () => {
             CUSTOM_INTERACTIONS.polling,
             1,
             '1.2.3',
-            {}
+            {},
           ),
         {
           timeout: WAIT_TIMEOUT,
-        }
+        },
       );
 
       // Reset mock to ensure it is triggered again with polling
@@ -1651,7 +1681,7 @@ describe('web-component', () => {
       nextMock.mockReturnValueOnce(
         generateSdkResponse({
           action: RESPONSE_ACTIONS.poll,
-        })
+        }),
       );
 
       // Click another button, which returns the same screen
@@ -1666,11 +1696,11 @@ describe('web-component', () => {
             CUSTOM_INTERACTIONS.polling,
             1,
             '1.2.3',
-            {}
+            {},
           ),
         {
           timeout: WAIT_TIMEOUT,
-        }
+        },
       );
     });
 
@@ -1682,7 +1712,7 @@ describe('web-component', () => {
       nextMock.mockReturnValue(
         generateSdkResponse({
           action: RESPONSE_ACTIONS.poll,
-        })
+        }),
       );
 
       pageContent = '<div data-type="polling">...</div><span>It works!</span>';
@@ -1704,12 +1734,12 @@ describe('web-component', () => {
         .mockReturnValueOnce(
           generateSdkResponse({
             action: RESPONSE_ACTIONS.poll,
-          })
+          }),
         )
         .mockReturnValueOnce(
           generateSdkResponse({
             status: 'completed',
-          })
+          }),
         );
 
       pageContent = '<div data-type="polling">...</div><span>It works!</span>';
@@ -1730,11 +1760,11 @@ describe('web-component', () => {
       await waitFor(
         () =>
           expect(onSuccess).toHaveBeenCalledWith(
-            expect.objectContaining({ detail: 'auth info' })
+            expect.objectContaining({ detail: 'auth info' }),
           ),
         {
           timeout: WAIT_TIMEOUT,
-        }
+        },
       );
 
       wcEle.removeEventListener('success', onSuccess);
@@ -1749,7 +1779,7 @@ describe('web-component', () => {
       startMock.mockReturnValueOnce(generateSdkResponse());
       localStorage.setItem(
         DESCOPE_LAST_AUTH_LOCAL_STORAGE_KEY,
-        '{"authMethod":"otp"}'
+        '{"authMethod":"otp"}',
       );
       getLastUserLoginIdMock.mockReturnValue('abc');
 
@@ -1785,7 +1815,7 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
     });
 
@@ -1793,7 +1823,7 @@ describe('web-component', () => {
       startMock.mockReturnValueOnce(generateSdkResponse());
       localStorage.setItem(
         DESCOPE_LAST_AUTH_LOCAL_STORAGE_KEY,
-        '{"authMethod":"otp"}'
+        '{"authMethod":"otp"}',
       );
 
       configContent = {
@@ -1828,7 +1858,7 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
     });
 
@@ -1836,7 +1866,7 @@ describe('web-component', () => {
       startMock.mockReturnValueOnce(generateSdkResponse());
       localStorage.setItem(
         DESCOPE_LAST_AUTH_LOCAL_STORAGE_KEY,
-        '{"authMethod":"otp"}'
+        '{"authMethod":"otp"}',
       );
       getLastUserLoginIdMock.mockReturnValue('abc');
 
@@ -1883,6 +1913,7 @@ describe('web-component', () => {
             samlIdpStateId: null,
             samlIdpUsername: null,
             ssoAppId: null,
+            client: {},
             redirectAuth: undefined,
             preview: false,
             tenant: undefined,
@@ -1891,15 +1922,15 @@ describe('web-component', () => {
           'interactionId',
           1,
           '1.2.3',
-          { origin: 'http://localhost' }
-        )
+          { origin: 'http://localhost' },
+        ),
       );
     });
     it('Should call start with code and idpInitiated when idpInitiated condition is met', async () => {
       window.location.search = `?${URL_CODE_PARAM_NAME}=code1`;
       localStorage.setItem(
         DESCOPE_LAST_AUTH_LOCAL_STORAGE_KEY,
-        '{"authMethod":"otp"}'
+        '{"authMethod":"otp"}',
       );
       getLastUserLoginIdMock.mockReturnValue('abc');
       configContent = {
@@ -1934,6 +1965,7 @@ describe('web-component', () => {
             ssoAppId: null,
             redirectAuth: undefined,
             tenant: undefined,
+            client: {},
             lastAuth: { authMethod: 'otp' },
           },
           undefined,
@@ -1943,8 +1975,8 @@ describe('web-component', () => {
           {
             exchangeCode: 'code1',
             idpInitiated: true,
-          }
-        )
+          },
+        ),
       );
     });
 
@@ -1981,7 +2013,7 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
     });
 
@@ -1989,7 +2021,7 @@ describe('web-component', () => {
       window.location.search = `?${URL_TOKEN_PARAM_NAME}=code1`;
       localStorage.setItem(
         DESCOPE_LAST_AUTH_LOCAL_STORAGE_KEY,
-        '{"authMethod":"otp"}'
+        '{"authMethod":"otp"}',
       );
       getLastUserLoginIdMock.mockReturnValue('abc');
       configContent = {
@@ -2024,6 +2056,7 @@ describe('web-component', () => {
             samlIdpStateId: null,
             samlIdpUsername: null,
             ssoAppId: null,
+            client: {},
           },
           undefined,
           '',
@@ -2031,8 +2064,8 @@ describe('web-component', () => {
           undefined,
           {
             token: 'code1',
-          }
-        )
+          },
+        ),
       );
     });
 
@@ -2068,7 +2101,7 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
     });
 
@@ -2093,6 +2126,7 @@ describe('web-component', () => {
             samlIdpStateId: null,
             samlIdpUsername: null,
             ssoAppId: null,
+            client: {},
             redirectAuth: { callbackUrl: callback, codeChallenge: challenge },
             tenant: undefined,
             lastAuth: {},
@@ -2101,8 +2135,8 @@ describe('web-component', () => {
           '',
           0,
           '1.2.3',
-          undefined
-        )
+          {},
+        ),
       );
       await waitFor(() => screen.findByShadowText('It works!'), {
         timeout: WAIT_TIMEOUT,
@@ -2134,13 +2168,14 @@ describe('web-component', () => {
             samlIdpStateId: null,
             samlIdpUsername: null,
             ssoAppId: null,
+            client: {},
           },
           undefined,
           '',
           0,
           '1.2.3',
-          { token }
-        )
+          { token },
+        ),
       );
       await waitFor(() => screen.findByShadowText('It works!'), {
         timeout: WAIT_TIMEOUT,
@@ -2167,6 +2202,7 @@ describe('web-component', () => {
             samlIdpStateId: null,
             samlIdpUsername: null,
             ssoAppId: null,
+            client: {},
             tenant: undefined,
             redirectAuth: undefined,
             lastAuth: {},
@@ -2175,8 +2211,8 @@ describe('web-component', () => {
           '',
           0,
           '1.2.3',
-          undefined
-        )
+          {},
+        ),
       );
       await waitFor(() => screen.findByShadowText('It works!'), {
         timeout: WAIT_TIMEOUT,
@@ -2292,6 +2328,7 @@ describe('web-component', () => {
             samlIdpStateId: 'abcdefgh',
             samlIdpUsername: null,
             ssoAppId: null,
+            client: {},
             tenant: undefined,
             redirectAuth: undefined,
             lastAuth: {},
@@ -2300,8 +2337,8 @@ describe('web-component', () => {
           '',
           0,
           '1.2.3',
-          undefined
-        )
+          {},
+        ),
       );
       await waitFor(() => screen.getByShadowText('It works!'), {
         timeout: WAIT_TIMEOUT,
@@ -2330,6 +2367,7 @@ describe('web-component', () => {
             samlIdpStateId: 'abcdefgh',
             samlIdpUsername: 'dummyUser',
             ssoAppId: null,
+            client: {},
             tenant: undefined,
             redirectAuth: undefined,
             lastAuth: {},
@@ -2338,8 +2376,8 @@ describe('web-component', () => {
           '',
           0,
           '1.2.3',
-          undefined
-        )
+          {},
+        ),
       );
       await waitFor(() => screen.getByShadowText('It works!'), {
         timeout: WAIT_TIMEOUT,
@@ -2399,13 +2437,14 @@ describe('web-component', () => {
             tenant: undefined,
             redirectAuth: undefined,
             lastAuth: {},
+            client: {},
           },
           undefined,
           '',
           0,
           '1.2.3',
-          undefined
-        )
+          {},
+        ),
       );
       await waitFor(() => screen.getByShadowText('It works!'), {
         timeout: WAIT_TIMEOUT,
@@ -2448,6 +2487,7 @@ describe('web-component', () => {
           samlIdpStateId: null,
           samlIdpUsername: null,
           ssoAppId: null,
+          client: {},
           redirectAuth: undefined,
           tenant: undefined,
           lastAuth: {},
@@ -2459,8 +2499,8 @@ describe('web-component', () => {
         {
           exchangeCode: 'code1',
           idpInitiated: true,
-        }
-      )
+        },
+      ),
     );
   });
 
@@ -2513,6 +2553,7 @@ describe('web-component', () => {
           redirectAuth: undefined,
           tenant: undefined,
           lastAuth: {},
+          client: {},
         },
         undefined,
         '',
@@ -2521,8 +2562,8 @@ describe('web-component', () => {
         {
           exchangeCode: 'code1',
           idpInitiated: true,
-        }
-      )
+        },
+      ),
     );
   });
 
@@ -2530,7 +2571,7 @@ describe('web-component', () => {
     startMock.mockReturnValueOnce(generateSdkResponse());
     localStorage.setItem(
       DESCOPE_LAST_AUTH_LOCAL_STORAGE_KEY,
-      '{"authMethod":"otp"}'
+      '{"authMethod":"otp"}',
     );
     getLastUserLoginIdMock.mockReturnValue('abc');
 
@@ -2580,14 +2621,14 @@ describe('web-component', () => {
 
     expect(fetchMock).toHaveBeenCalledWith(
       expect.stringMatching(htmlUrlPathRegex),
-      expect.any(Object)
+      expect.any(Object),
     );
   });
   it('Should fetch else screen when else is met', async () => {
     startMock.mockReturnValueOnce(generateSdkResponse());
     localStorage.setItem(
       DESCOPE_LAST_AUTH_LOCAL_STORAGE_KEY,
-      '{"authMethod":"otp"}'
+      '{"authMethod":"otp"}',
     );
     getLastUserLoginIdMock.mockReturnValue('');
 
@@ -2643,7 +2684,7 @@ describe('web-component', () => {
 
     expect(fetchMock).toHaveBeenCalledWith(
       expect.stringMatching(htmlUrlPathRegex),
-      expect.any(Object)
+      expect.any(Object),
     );
   });
 
@@ -2654,7 +2695,7 @@ describe('web-component', () => {
       generateSdkResponse({
         ok: true,
         status: 'completed',
-      })
+      }),
     );
 
     document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="otpSignInEmail" project-id=1></descope-wc>`;
@@ -2668,9 +2709,9 @@ describe('web-component', () => {
     await waitFor(
       () =>
         expect(onSuccess).toHaveBeenCalledWith(
-          expect.objectContaining({ detail: 'auth info' })
+          expect.objectContaining({ detail: 'auth info' }),
         ),
-      { timeout: WAIT_TIMEOUT }
+      { timeout: WAIT_TIMEOUT },
     );
 
     wcEle.removeEventListener('success', onSuccess);
@@ -2698,17 +2739,17 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(themeUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(configUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
     });
 
@@ -2744,20 +2785,20 @@ describe('web-component', () => {
 
         expect(fetchMock).toHaveBeenCalledWith(
           expect.stringMatching(htmlUrlPathRegex),
-          expect.any(Object)
+          expect.any(Object),
         );
 
         expect(fetchMock).toHaveBeenCalledWith(
           expect.stringMatching(themeUrlPathRegex),
-          expect.any(Object)
+          expect.any(Object),
         );
 
         expect(fetchMock).toHaveBeenCalledWith(
           expect.stringMatching(configUrlPathRegex),
-          expect.any(Object)
+          expect.any(Object),
         );
       },
-      WAIT_TIMEOUT
+      WAIT_TIMEOUT,
     );
 
     it('should fetch the data from the correct path when locale provided and not part of target locales', async () => {
@@ -2790,17 +2831,17 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(themeUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(configUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
     });
 
@@ -2839,17 +2880,17 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(themeUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(configUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       Object.defineProperty(navigator, 'language', {
@@ -2892,17 +2933,17 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(themeUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(configUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       Object.defineProperty(navigator, 'language', {
@@ -2946,17 +2987,17 @@ describe('web-component', () => {
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(themeUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(configUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       Object.defineProperty(navigator, 'language', {
@@ -3005,29 +3046,29 @@ describe('web-component', () => {
 
       const htmlUrlPathRegex = new RegExp(`//[^/]+${expectedHtmlPath}$`);
       const htmlUrlFallbackPathRegex = new RegExp(
-        `//[^/]+${expectedHtmlFallbackPath}$`
+        `//[^/]+${expectedHtmlFallbackPath}$`,
       );
       const themeUrlPathRegex = new RegExp(`//[^/]+${expectedThemePath}$`);
       const configUrlPathRegex = new RegExp(`//[^/]+${expectedConfigPath}$`);
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(htmlUrlFallbackPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(themeUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       expect(fetchMock).toHaveBeenCalledWith(
         expect.stringMatching(configUrlPathRegex),
-        expect.any(Object)
+        expect.any(Object),
       );
 
       Object.defineProperty(navigator, 'language', {
@@ -3049,7 +3090,7 @@ describe('web-component', () => {
           samlIdpResponseUrl: samlUrl,
           samlIdpResponseSamlResponse: 'saml-response-dummy-value',
           samlIdpResponseRelayState: 'saml-relay-state-dummy-value',
-        })
+        }),
       );
 
       const mockSubmitForm = jest.spyOn(helpers, 'submitForm');
@@ -3068,14 +3109,14 @@ describe('web-component', () => {
         },
         {
           timeout: 6000,
-        }
+        },
       )) as HTMLFormElement;
 
       expect(form).toBeInTheDocument();
 
       // validate inputs exist
       const inputSamlResponse = document.querySelector(
-        `form[action="${samlUrl}"] input[role="saml-response"]`
+        `form[action="${samlUrl}"] input[role="saml-response"]`,
       );
       expect(inputSamlResponse).toBeInTheDocument();
       expect(inputSamlResponse).not.toBeVisible();
@@ -3083,7 +3124,7 @@ describe('web-component', () => {
 
       // validate inputs are hidden
       const inputSamlRelayState = document.querySelector(
-        `form[action="${samlUrl}"] input[role="saml-relay-state"]`
+        `form[action="${samlUrl}"] input[role="saml-relay-state"]`,
       );
       expect(inputSamlRelayState).toBeInTheDocument();
       expect(inputSamlRelayState).not.toBeVisible();
@@ -3093,7 +3134,7 @@ describe('web-component', () => {
         () => {
           expect(mockSubmitForm).toHaveBeenCalledTimes(1);
         },
-        { timeout: 6000 }
+        { timeout: 6000 },
       );
     });
 
@@ -3102,7 +3143,7 @@ describe('web-component', () => {
         generateSdkResponse({
           ok: true,
           executionId: 'e1',
-        })
+        }),
       );
       nextMock.mockReturnValueOnce(generateSdkResponse({ screenId: '1' }));
 
@@ -3123,7 +3164,7 @@ describe('web-component', () => {
         () => screen.findAllByShadowDisplayValue(samlIdpEmailAddress),
         {
           timeout: 6000,
-        }
+        },
       );
 
       expect(inputs.length).toBe(2);
@@ -3146,8 +3187,8 @@ describe('web-component', () => {
         expect(errorSpy).toHaveBeenCalledWith(
           'Cannot load DescopeUI',
           expect.any(String),
-          expect.any(Error)
-        )
+          expect.any(Error),
+        ),
       );
     });
     it('should try to load all descope component on the page', async () => {
@@ -3166,9 +3207,9 @@ describe('web-component', () => {
       await waitFor(
         () =>
           Object.keys(globalThis.DescopeUI).forEach((key) =>
-            expect(globalThis.DescopeUI[key]).toHaveBeenCalled()
+            expect(globalThis.DescopeUI[key]).toHaveBeenCalled(),
           ),
-        { timeout: WAIT_TIMEOUT }
+        { timeout: WAIT_TIMEOUT },
       );
     });
     it('should log an error if descope component is missing', async () => {
@@ -3186,9 +3227,9 @@ describe('web-component', () => {
           expect(errorSpy).toHaveBeenCalledWith(
             'Cannot load UI component "descope-button1"',
             expect.any(String),
-            expect.any(Error)
+            expect.any(Error),
           ),
-        { timeout: 5100 }
+        { timeout: 5100 },
       );
     });
     it('should not load components which are already loaded', async () => {
@@ -3199,7 +3240,7 @@ describe('web-component', () => {
 
       customElements.define(
         'descope-test-button',
-        class extends HTMLElement {}
+        class extends HTMLElement {},
       );
 
       const DescopeUI = { 'descope-test-button': jest.fn() };